requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
alsa-lib=1.2.12=h4ab18f5_0
asttokens=2.4.1=pyhd8ed1ab_0
attr=2.5.1=h166bdaf_1
backcall=0.2.0=pyh9f0ad1d_0
blas=1.0=mkl
bottleneck=1.3.7=py38ha9d4c09_0
brotli=1.0.9=h5eee18b_8
brotli-bin=1.0.9=h5eee18b_8
brotli-python=1.0.9=py38h6a678d5_8
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.9.24=h06a4308_0
cairo=1.18.0=h3faef2a_0
certifi=2024.8.30=py38h06a4308_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.0.5=py38hdb19cb5_0
cuda-cudart=12.4.127=0
cuda-cupti=12.4.127=0
cuda-libraries=12.4.0=0
cuda-nvrtc=12.4.127=0
cuda-nvtx=12.4.127=0
cuda-opencl=12.6.77=0
cuda-runtime=12.4.0=0
cuda-version=12.6=3
cycler=0.11.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
debugpy=1.6.7=py38h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
et_xmlfile=1.1.0=py38h06a4308_0
executing=2.1.0=pyhd8ed1ab_0
expat=2.6.3=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py38h06a4308_0
font-ttf-dejavu-sans-mono=2.37=hd3eb1b0_0
font-ttf-inconsolata=2.001=hcb22688_0
font-ttf-source-code-pro=2.030=hd3eb1b0_0
font-ttf-ubuntu=0.83=h8b1ccd4_0
fontconfig=2.14.2=h14ed4e7_0
fonts-anaconda=1=h8fa9717_0
fonts-conda-ecosystem=1=hd3eb1b0_0
fonttools=4.51.0=py38h5eee18b_0
freetype=2.12.1=h4a9f257_0
gettext=0.22.5=he02047a_3
gettext-tools=0.22.5=he02047a_3
glib=2.80.2=hf974151_0
glib-tools=2.80.2=hb6ce0ca_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py38heeb90bb_0
gnutls=3.6.15=he1e5248_0
graphite2=1.3.14=h295c915_1
gst-plugins-base=1.24.4=h9ad1361_0
gstreamer=1.24.4=haf2f30d_0
harfbuzz=8.5.0=hfac3d4d_0
icu=73.2=h59595ed_0
idna=3.7=py38h06a4308_0
importlib-metadata=8.5.0=pyha770c72_0
importlib_resources=6.4.0=py38h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.29.5=pyh3099207_0
ipython=8.12.2=pyh41d4057_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.4=py38h06a4308_0
joblib=1.4.2=py38h06a4308_0
jupyter_client=8.6.3=pyhd8ed1ab_0
jupyter_core=5.7.2=pyh31011fe_1
kiwisolver=1.4.4=py38h6a678d5_0
krb5=1.21.3=h143b758_0
lame=3.100=h7b6447c_0
lcms2=2.16=hb7c19ff_0
ld_impl_linux-64=2.40=h12ee557_0
lerc=4.0.0=h27087fc_0
libasprintf=0.22.5=he8f35ee_3
libasprintf-devel=0.22.5=he8f35ee_3
libbrotlicommon=1.0.9=h5eee18b_8
libbrotlidec=1.0.9=h5eee18b_8
libbrotlienc=1.0.9=h5eee18b_8
libcap=2.69=h0f662aa_0
libclang-cpp15=15.0.7=default_h127d8a8_5
libclang13=18.1.4=default_h5d6823c_0
libcublas=12.4.2.65=0
libcufft=11.2.0.44=0
libcufile=1.11.1.6=0
libcups=2.3.3=h4637d8d_4
libcurand=10.3.7.77=0
libcusolver=11.6.0.99=0
libcusparse=12.3.0.142=0
libdeflate=1.20=hd590300_0
libedit=3.1.20230828=h5eee18b_0
libevent=2.1.12=hdbd6064_1
libexpat=2.6.3=h5888daf_0
libffi=3.4.4=h6a678d5_1
libflac=1.4.3=h59595ed_0
libgcc=14.1.0=h77fa898_1
libgcc-ng=14.1.0=h69a702a_1
libgcrypt=1.11.0=h4ab18f5_1
libgettextpo=0.22.5=he02047a_3
libgettextpo-devel=0.22.5=he02047a_3
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libglib=2.80.2=hf974151_0
libgomp=14.1.0=h77fa898_1
libgpg-error=1.50=h4f305b6_0
libiconv=1.17=hd590300_2
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=3.0.3=h5eee18b_0
libllvm15=15.0.7=hb3ce162_4
libllvm18=18.1.4=h2448989_0
libnpp=12.2.5.2=0
libnvfatbin=12.6.77=0
libnvjitlink=12.4.99=0
libnvjpeg=12.3.1.89=0
libogg=1.3.5=h27cfd23_1
libopus=1.3.1=h7b6447c_0
libpng=1.6.43=h2797004_0
libpq=16.4=h2d7952a_1
libsndfile=1.2.2=hc60ed4a_1
libsodium=1.0.20=h4ab18f5_0
libsqlite=3.46.0=hde9e2c9_0
libstdcxx=14.1.0=hc0a3c3a_1
libstdcxx-ng=14.1.0=h4852527_1
libsystemd0=255=h3516f8a_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.6.0=h1dd3fc0_3
libunistring=0.9.10=h27cfd23_0
libuuid=2.38.1=h0b41bf4_0
libvorbis=1.3.7=h7b6447c_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.7.0=h662e7e4_0
libxml2=2.13.1=hfdd30dd_2
libzlib=1.2.13=h4ab18f5_6
llvm-openmp=14.0.6=h9e868ea_0
llvmlite=0.41.1=pypi_0
lz4-c=1.9.4=h6a678d5_1
markupsafe=2.1.3=py38h5eee18b_0
matplotlib=3.7.2=py38h06a4308_0
matplotlib-base=3.7.2=py38h1128e8f_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.8=py38h5eee18b_0
mkl_random=1.2.4=py38hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpg123=1.32.6=h59595ed_0
mpmath=1.3.0=py38h06a4308_0
mysql-common=8.3.0=hf1915f5_4
mysql-libs=8.3.0=hca2cd23_4
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py38h06a4308_0
nspr=4.35=h6a678d5_0
nss=3.100=hca3bf56_0
numba=0.58.1=pypi_0
numexpr=2.8.4=py38hc78ab66_1
numpy=1.24.3=py38hf6e8229_1
numpy-base=1.24.3=py38h060ed82_1
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=he7f1fd0_0
openpyxl=3.1.5=py38h5eee18b_0
openssl=3.3.2=hb9d3cd8_0
packaging=24.1=pyhd8ed1ab_0
pandas=2.0.3=py38h1128e8f_0
parso=0.8.4=pyhd8ed1ab_0
pcre2=10.43=hcad00b1_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.3.0=py38h9e66945_0
pip=24.2=py38h06a4308_0
pixman=0.43.2=h59595ed_0
platformdirs=4.3.6=pyhd8ed1ab_0
ply=3.11=py38_0
pooch=1.7.0=py38h06a4308_0
prompt-toolkit=3.0.48=pyha770c72_0
prompt_toolkit=3.0.48=hd8ed1ab_0
psutil=6.0.0=py38hfb59056_0
ptyprocess=0.7.0=pyhd3deb0d_0
pulseaudio-client=17.0=hb77b528_0
pure_eval=0.2.3=pyhd8ed1ab_0
pygments=2.18.0=pyhd8ed1ab_0
pynndescent=0.5.13=pypi_0
pyparsing=3.0.9=py38h06a4308_0
pyqt=5.15.10=py38h6a678d5_0
pyqt5-sip=12.13.0=py38h5eee18b_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.19=h955ad1f_0
python-dateutil=2.9.0=pyhd8ed1ab_0
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytorch=2.4.1=py3.8_cuda12.4_cudnn9.1.0_0
pytorch-cuda=12.4=hc786d27_6
pytorch-mutex=1.0=cuda
pytz=2024.1=py38h06a4308_0
pyyaml=6.0.1=py38h5eee18b_0
pyzmq=25.1.2=py38h6a678d5_0
qt-main=5.15.8=hc9dc06e_21
readline=8.2=h5eee18b_0
requests=2.32.3=py38h06a4308_0
scikit-learn=1.3.0=py38h1128e8f_1
scipy=1.10.1=py38hf6e8229_1
setuptools=75.1.0=py38h06a4308_0
sip=6.7.12=py38h6a678d5_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.45.3=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.13.2=py38h06a4308_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=3.5.0=py38h2f386ee_0
tk=8.6.14=h39e8969_0
tomli=2.0.1=py38h06a4308_0
torchaudio=2.4.1=py38_cu124
torchtriton=3.0.0=py38
torchvision=0.19.1=py38_cu124
tornado=6.4.1=py38hfb59056_0
tqdm=4.66.5=pypi_0
traitlets=5.14.3=pyhd8ed1ab_0
typing_extensions=4.12.2=pyha770c72_0
umap-learn=0.5.6=pypi_0
unicodedata2=15.1.0=py38h5eee18b_0
urllib3=2.2.2=py38h06a4308_0
wcwidth=0.2.13=pyhd8ed1ab_0
wheel=0.44.0=py38h06a4308_0
xcb-util=0.4.0=hd590300_1
xcb-util-image=0.4.0=h8ee46fc_1
xcb-util-keysyms=0.4.0=h8ee46fc_1
xcb-util-renderutil=0.3.9=hd590300_1
xcb-util-wm=0.4.1=h8ee46fc_1
xkeyboard-config=2.42=h4ab18f5_0
xorg-kbproto=1.0.7=hb9d3cd8_1003
xorg-libice=1.1.1=hb9d3cd8_1
xorg-libsm=1.2.4=h7391055_0
xorg-libx11=1.8.9=h8ee46fc_0
xorg-libxau=1.0.11=hb9d3cd8_1
xorg-libxext=1.3.4=h0b41bf4_2
xorg-libxrender=0.9.11=hd590300_0
xorg-renderproto=0.11.1=hb9d3cd8_1003
xorg-xextproto=7.3.0=hb9d3cd8_1004
xorg-xf86vidmodeproto=2.3.1=hb9d3cd8_1003
xorg-xproto=7.0.31=h27cfd23_1007
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
zeromq=4.3.5=ha4adb4c_5
zipp=3.20.2=pyhd8ed1ab_0
zlib=1.2.13=h4ab18f5_6
zstd=1.5.5=hc292b87_2
